mOS: Kernel core
Co-authored-by: Tom Musta <tom.musta@intel.com>
Co-authored-by: Rolf Riesen <rolf.riesen@intel.com>
Co-authored-by: John Attinella <john.e.attinella@intel.com>
Co-authored-by: Sharath Kumar Bhat <sharath.k.bhat@intel.com>
Co-authored-by: Andrew Tauferner <andrew.t.tauferner@intel.com>
Co-authored-by: Balazs Gerofi <balazs.gerofi@intel.com>
Signed-off-by: John Attinella <john.e.attinella@intel.com>

Add example configuration file: config.mos

mos-core: doc updates and ras support

Update README with v1.0 information

mOS kernel GPU isolation

Track GPU resources in the kernel.

These are the base mOS kernel updates to support isolation of
GPU resources. New mOS sysfs files have been introduced:

/sys/kernel/mOS/lwkgpus
/sys/kernel/mOS/lwkgpus_mask
These files contain the list of GPU devices that are designated for use
by the LWK. Unlike the lwkcpus file, this file is set by lwkctl from
user-state after it calls into the OneAPI Level0 interfaces to find the
available GPU devices. Also unlike CPUs, these devices are not fenced
off from Linux within the kernel.

/sys/kernel/mOS/lwkgpus_reserved
/sys/kernel/mOS/lwkgpus_reserved_mask
These files indicate whether one or more processes have reserved this
device. Unlike the lwkcpus_reserved file, more than one process may have
reserved this device at the same time.

/sys/kernel/mOS/lwkgpus_usage_count
This file shows how may processes have a device reserved. When reading
the file, a comma-separated usage count is returned for each designated
GPU. For example if there are 6 devices, the file contents could be
0,0,0,0,0,0

/sys/kernel/mOS/lwkgpus_numa
This file returns the numa domain associated with the GPU devices. When
reading this file, a comma-separated numa domain id is returned for each
designated GPU. For example if there are 6 devices, the file contents
could be: 0,0,0,1,1,1

/sys/kernel/mOS/lwkgpus_request
This file is a write-only file, similiar to lwkcpus_request. When YOD
decides to reserve GPU devices for a process, it will write a list of
the devices to this file. Unlike the lwkcpus_request, the operation will
succeed even if another process has a reservation. When the write
operation is done, the usage_count for each GPU device will be
incremented. Also the mask of reserved GPUs will be set into the mOS
process structure within the kernel. When that lwk process exits, it
will decrement the usage counts for the GPU devices that it had
reserved. If the usage count decrements to zero, the corresponding
reserved indicator in the lwkgpus_reserved file will be set to zero.

The gpmask interfaces have been enhanced to provide additional function.

mos-core: fix reserved LWK CPUs after process exit

When a yod call conflicts with the CPUs already reserved in an LWK
process, the bits in the lwkcpus_reserved_map bitmask associated with
the LWK CPUs of the running process may get cleared. Later when that
process ends, it does an XOR of its process-scoped bit mask with the
systems lwkcpus_reserved_map. Since the previously failed YOD has
reset this bits, the XOR operation that was meant to turn off the bits
in the lwkcpus_reserved_map actually ends up turning the bits on. The
process then exits, thinking everything is OK. These reserved bits will
prevent the lwk partition from being deleted or modified. It will also
limit the resources available to any subsequent jobs that run in the
current lwk partition. The only "supported" way to recover from this
problem is a system reboot. There is a less drastic way to recover which
involves writing NULL to the /sys/kernel/mOS/lwkcpus_reserved file under
sudo. A manifestation of this problem is in HPCS-5993.

mos-core: Option to inline offlining LWK CPUs with bringup

mos-core: Allow selected interrupts on LWK CPUs

mos-core: Offline CPUs together when inline offlining CPUs

mos-core: change warning RAS to informational msg

When the balancer attempts to move an irq to and LWK CPU and that
IRQ was deemed not to be allowed on an LWK CPU, do not generate
a warning MOS RAS message. Instead generate an informational message.
This is because we now allow the balancer to attempt to move IRQs to
LWK CPUs because we will allow IRQs from certain devices to exist on
LWK CPUs.

mos-core: Remove option to select inline offlining

mos-core: bump version from 0.9 to 1.0

mos-core: Coverity fixes