mOS-tools: yod, lwkreset, lwkctl
Co-authored-by: Tom Musta <tom.musta@intel.com>
Co-authored-by: Sharath Kumar Bhat <sharath.k.bhat@intel.com>
Co-authored-by: John Attinella <john.e.attinella@intel.com>
Co-authored-by: Andrew Tauferner <andrew.t.tauferner@intel.com>

Signed-off-by: John Attinella <john.e.attinella@intel.com>

GPU isolation in lwkctl and yod

This commit contains new function to search for available GPU devices
and designate those devices for use within an LWK partition. The OneAPI
LevelZero interfaces are used, along with other sysfs files to determine
the list of devices and their associated NUMA domains. The device and
NUMA domain information is writting in the /sys/kernel/mOS/lwkgpus and
/sys/mOS/lwkgpus_numa files. The usage counts located in the
/sys/kernel/mOS/lwkgpus_usage_count file are reset to zero. All GPUs
that are seen by the LevelZero interface are given to the LWK partition.
Note that Linux also retains access to these devices.

Since the management of GPUs has dependency on OneAPI LevelZero, and
since all enviroments that use mOS may not have access to this package,
a config setting has been introduced so that mOS can be built when the
OneAPI LevelZero package is not available. The kernel config value is
the following:
CONFIG_MOS_ONEAPI_LEVELZERO=y
This config setting must be included in the ".config" file  when
building mOS if management of GPU devices is desired. If not provided,
GPU devices will still be available to the LWK partitions, but any
isolation would need to be done by the caller setting the
ZE_AFFINITY_MASK prior to calling YOD or from within the process.

At support for tile granularity when performing resource management
of GPU devices. When reserving resources for an LWK process, the
'-R' option will divide resources at tile granularity. Entire devices
will be used if the fraction of resources requested allows it. When using
the '--gpus' option and providing a number or fraction, it will be
assumed that device granularity is being requested. A new option is now
supported, '--gpu-tiles', that can be used if the user wishes to
allocate a specific number of tiles. Also the internal environment
variable that should be set when using PVC A-step level hardware is now
required to be set when creating the LWK partition and not when 'yod'
is being called. Also the name was changed to MOS_PVC_ASTEP.

The layout of the bits in the lwkgpus_mask now contains information on
the subdevices (tiles). Each NIBBLE in the bitmask represents a GPU
device, and each bit within the nibble represents a tile. With the
current configuration, a maximum of 16 GPU devices with a maximum of 4
tiles each is supported. Using defines in the newly introduced
mOS/tools/include/gpumask.h file, this compile-time configuration can be
modify to support any combination of MAX tiles and MAX devices as long
as the product of the two does not exceed 64. The 64 limit is due to the
limit currently imposed by the mOS kernel.

This is an enhancement to the recently integrated support
of GPUs. If the ZE_AFFINITY_MASK environment variable is set
on entry to YOD, and if the --gpus or --gpu_tiles options are
not specified, then the ZE_AFFINITY_MASK will drive the core and
numa node selection for the current LWK process. For example, if
ZE_AFFINITY_MASK=5 and if device 5 is in numa node 1, then mOS will
choose cores from node 1 and reserve memory from node 1. This
assumes that the --resources and --cores options are requesting a
number of cores that can be contained within that numa node. If the
number of cores requested is greater than the number of cores within
scope of the numa nodes defined by the selected GPU devices, then
cores from outside of the GPU numa nodes will be included and no
error will be indicated.

yod: add YOD_ONLY_TILE_0 environment variable

Add YOD_ONLY_TILE_0 environment variable. When this is
set to a value of 1, then only tile 0 of each GPU device
will be considered for reservation in the LWK process. The
primary purpose of this change is to support the PVC A0 stepping
in which both tiles cannot be reliably used at the same time on
the GPU device.

yod: Add GPU support to the resource map file

Add the --gpus and --gpu-tiles options to the resource
map file.

YOD: support MPI value for cores, gpus, memory

The --resource option in YOD supports a value of MPI. This uses
MPI environment variables to generate the fractional resources to
allocate for cores, GPUs, and memory. This change allows the MPI
value to be use on the --cores, --gpus, --gpu_tiles, and --mem options.
This flexibility can be useful when we want to explicitly select a
specific number of one type of resource but want to divide the other
resources based on the number of ranks specified in the mpirun
parameters.
For example:
yod -C 2 -M mpi -G mpi myApplication

lwkctl: option to inline offlining LWK CPUs with bringup

lwkctl: Allow selected interrupts on LWK CPUs

lwkctl: Offline sibling CPUs together when inline offlining CPUs

mOS-tools: remove setting of IRQ_BANNED_CPUS

Remove setting the environment variable in the IRQ balancer that
would restrict the balancer from using LWK CPUs.

lwkctl: remove option to select inline offlining

lwkctl: add i915 and HFI to the driver list

Allow interrupt vectors allocated by the i915 and HFI drivers to exist
on LWK CPUs.

Fix compiler warnings in lwkctl and yod

mos-tools: Coverity fixes